iw/products/InteractiveWall/assets/database/requirements.txt
django==1.10
djangorestframework
markdown
pep8